+++ packages/adminui-vue/components/controls/mixins/alsMixin.js
@@ -1,15 +1,15 @@
 /**
  * Mixin provides the following functionality:
  * methods:
- *   isReadOnlyByALS - check that exist rule for disable field update
- *   isRequiredByALS - check that exist rule to make field mandatory/required
+ *   $_isReadOnlyByALS - check that exist rule for disable field update
+ *   $_isRequiredByALS - check that exist rule to make field mandatory/required
  * To make this mixin work use it on form as:
  * const alsMixin = require('@adminui-vue/components/controls/mixins/alsMixin')
  * mixins: ['alsMixin']
  */
 module.exports = {
   methods: {
-    isReadOnlyByALS (attributeName) {
+    $_isReadOnlyByALS (attributeName) {
       if (this.$store.state.alsInfo) {
         if (this.$store.state.alsInfo[attributeName]) {
           return this.$store.state.alsInfo[attributeName].indexOf('U') === -1 // if exist als rule check for possible update
@@ -17,7 +17,7 @@ module.exports = {
       } else return false // if not exist als mixin
     },
 
-    isRequiredByALS (attributeName) {
+    $_isRequiredByALS (attributeName) {
       if (this.$store.state.alsInfo && this.$store.state.alsInfo[attributeName]) {
         return this.$store.state.alsInfo[attributeName].indexOf('M') > -1
       } else return false