+++ packages/adminui-vue/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - The `processing` vuex module now has `SET_ALS_INFO` mutation.
    In case als mixin assigned to entity `processing.load` adds information about als
    into `alsInfo` object in vuex state.
+ - Added `alsMixin` and als support to `UAutoField.vue`
    
 ### Changed